CHG0152028: refactor: change usr passwr in Data400Controller; move some APIs to custom controller
